+++ lib/i18n_proofreading/engine.rb
@@ -22,5 +22,14 @@ module I18nProofreading
         include I18nProofreading::TagHelper
       end
     end
+
+    initializer 'i18n_proofreading.routing' do
+      ActionDispatch::Routing::Mapper.include(Module.new do
+        def mount_i18n_proofreading(at: I18nProofreading.config.mount_path, **options)
+          I18nProofreading.config.mount_path = at
+          mount I18nProofreading::Engine, at:, **options
+        end
+      end)
+    end
   end
 end