+++ test/dummy/config/routes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount I18nProofreading::Engine => '/i18n_proofreading'
+  mount_i18n_proofreading at: '/i18n_proofreading'
   get 'sample', to: 'sample#show'
 end